Extract ArrayColumnHelper from ArrayColumnFunctionReturnTypeExtension
